// src/common/enums/role.enum.ts
export enum UserRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  SCHOOL_ADMIN = 'SCHOOL_ADMIN',
  PRINCIPAL = 'PRINCIPAL',
  VICE_PRINCIPAL = 'VICE_PRINCIPAL',
  TEACHER = 'TEACHER',
  HOMEROOM_TEACHER = 'HOMEROOM_TEACHER',
  PARENT = 'PARENT',
  STUDENT = 'STUDENT',
  ACCOUNTANT = 'ACCOUNTANT',
  STAFF = 'STAFF',
}